Level Editing
Paved the road and moved a lot of things around.  Also made a modified version of the FPS character with the gun placed in a position that was a little more to my liking.  The original one is still there if you need it.  Also, I modified the tower model so that the player couldn't walk out of it, and has more vision on the paths below.