Level Design, New Models, Prefabs, & Textures
I did world building of the Level to near-completion using Sam and I's art assets.  Also made new models such as the bullet, mug, and citadel tower building for the city.  Modified some textures such as the ammo box to made it easier to see the text on the side.  Modified the FOV values on the Scope Zoom Script Slightly from 21 and 60 to 20 and 75.